Move InstalledLiveries tab content to subdirectory
