+++ src/components/HomeTabs/index.js
@@ -1,4 +1,4 @@
 export { default as AvailableLiveries } from './AvailableLiveries/AvailableLiveries';
-export { default as InstalledLiveries } from './InstalledLiveries';
+export { default as InstalledLiveries } from './InstalledLiveries/InstalledLiveries';
 export { default as Feed } from './Feed/Feed';
 export { default as Settings } from './Settings';
